main/backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# AI & ML
anthropic==0.18.1
openai==1.12.0
tiktoken==0.6.0

# Vector Store & RAG
chromadb==0.4.22
sentence-transformers==2.3.1

# File Handling
python-magic==0.4.27
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0

# WebSocket
websockets==12.0

# HTTP Client
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
